Warn when `RawMessageSerializer` reads an empty message
